Move form validation checks to separate file
Add a separate file for the form validation functions.
