+++ src/passari_web_ui/ui/forms.py
@@ -7,6 +7,11 @@ from passari_web_ui.db import db
 from passari_web_ui.ui.utils import get_available_object_count
 from passari_workflow.db.models import MuseumObject
 
+from .validators import (
+    object_ids_exist_check,
+    object_with_reason_exist_check
+)
+
 
 class EnqueueObjectsForm(FlaskForm):
     """
@@ -68,29 +73,6 @@ class MultipleObjectIDField(Field):
             self.data = []
 
 
-def object_ids_exist_check(form, field):
-    """
-    Check that all given object IDs exist
-    """
-    if not field.data:
-        raise ValidationError("No object ID was provided")
-
-    existing_object_ids = [
-        result[0] for result in
-        db.session.query(MuseumObject.id)
-        .filter(MuseumObject.id.in_(field.data))
-        .all()
-    ]
-
-    missing_object_ids = set(field.data) - set(existing_object_ids)
-
-    if missing_object_ids:
-        raise ValidationError(
-            f"Following objects don't exist: "
-            f"{', '.join([str(o) for o in sorted(missing_object_ids)])}"
-        )
-
-
 class FreezeObjectsForm(FlaskForm):
     """
     Form for freezing multiple objects with a single reason
@@ -106,21 +88,6 @@ class FreezeObjectsForm(FlaskForm):
         ),
         validators=[InputRequired()]
     )
-
-
-def object_with_reason_exist_check(form, field):
-    """
-    Check that at least one frozen object with the given reason exists
-    """
-    result = (
-        db.session.query(MuseumObject)
-        .filter(MuseumObject.frozen)
-        .filter(MuseumObject.freeze_reason == field.data)
-        .first()
-    )
-
-    if not result:
-        raise ValidationError("No objects with this reason were found.")
 
 
 class UnfreezeObjectsForm(FlaskForm):
